redesigned database design for longer and complex writes but easy and faster reads as GET request will be most frequent
