+++ models/mongoDB/forImport/answers.js
@@ -1,5 +1,4 @@
-// date must be a string in YYYY-mm-dd format;
-const { Answer } = require('./index');
+const { AnswerImport } = require('../index');
 
 const insertMany = async (answers) => {
   try {
@@ -16,7 +15,7 @@ const insertMany = async (answers) => {
       };
       return answerSchema;
     });
-    const bulkInserted = Answer.insertMany(
+    const bulkInserted = await AnswerImport.insertMany(
       answersToInsert,
       {
         limit: 10000, // 1000 seems to be the most optimal
@@ -30,26 +29,6 @@ const insertMany = async (answers) => {
   }
 };
 
-const findAllBy = async (questionId) => {
-  try {
-    const answersFound = await Answer.aggregate([
-      { $match: { question_id: Number(questionId) } },
-      {
-        $lookup: {
-          from: 'answersphotos',
-          localField: 'answer_id',
-          foreignField: 'answer_id',
-          as: 'photos',
-        },
-      },
-    ]);
-    return answersFound;
-  } catch (error) {
-    console.error(error);
-  }
-};
-
 module.exports = {
   insertMany,
-  findAllBy,
 };
