+++ server/controller/questions.js
@@ -1,6 +1,11 @@
-const getQuestions = (req, res) => {
+const { } = require('../../models/mongoDB/questions');
+
+const getQuestions = async (req, res) => {
   console.log(req.query);
   // req.query.product_id
+  const productId = req.query.product_id;
+  // const foundQuestions = await findAllBy(productId);
+  // console.log(foundQuestions[0].answers);
   // req.query.page
   // req.query.count
   res.send('getting all answers');
